refactor(tests): rename KSailStartCommandTests and consolidate with KSailStopCommandTests
