+++ tests/KSail.Tests/Commands/StartStop/KSailStartCommandTests.KSailStartHelp_SucceedsAndPrintsIntroductionAndHelp.verified.txt
